# backend/swagger.yml
swagger: "2.0"
info:
  description: |
    Custom API for pre-sales deal tracker

    The api has been designed arround three major resources: users, deals, and pitch requests
    Being the first version, designed by a junior, the api has room for improvements and best practices

    Route Specifications:
    - routes with admin are restricted for users with an admin role
    - routes with sales are restricted for users with a sales role
    - routes with no admin or sales are open endpoints for users with either role

    passwords can be refered to as a sub-resource in this version, because the passwords can be controlled by
    every user but also the admin has authority to reset every users password incase of a password loss.

    Setting default values for the requests objects can be tricky because of the flow of the poor design
  version: 1.0.0
  title: API Endpoints for Pre-sales Deal Tracker
host: "localhost:8080" #yet to be set or use localhost :)
basePath: /
...
schemes:
  - http
  - https
paths:
...
  /users/passwords:
    patch:
      # tags:
      #   - Update password
      summary: Uses data in request to update user password
      description: "Every user can update their passwords with this endpoint"
      operationId: updatePassword
      produces:
        - "application/json"
      parameters:
        - name: UpdatePasswordReq
          in: body
          description: holds the data need to update a users password
          required: true
          schema:
            $ref: "#/definitions/UpdatePasswordReq"
      responses:
        "204":
          description: successfull operation but updated resource is not returned
          schema:
            $ref: "#/definitions/successMsg"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/errorResp"
        "401":
          description: unauthorized access
          schema:
            $ref: "#/definitions/errorResp"
        "404":
          description: resource doesn't exist in database
          schema:
            $ref: "#/definitions/errorResp"
        "500":
          description: server error
          schema:
            $ref: "#/definitions/errorResp"
...
definitions:
  errorResp:
    type: object
    properties:
      error:
        type: string
  successMsg:
    type: object
    properties:
      message:
        type: string
...
  UpdatePasswordReq:
    type: object
    properties:
      old_password:
        type: string
      password_update:
        type: string
      user_id:
        type: integer
...